finish

+++ carpark/src/main/java/com/example/carpark/repo/BookingOfficeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookingOfficeRepository extends JpaRepository<BookingOffice, Integer> {
+public interface BookingOfficeRepository extends JpaRepository<BookingOffice, Long> {
 }
